rename sdk->trackjs. fix tests deprecation warnings.
