+++ e2e/core/tsconfig.json
@@ -3,10 +3,7 @@
   "compilerOptions": {
     "rootDir": "./",
     "noEmit": true,
-    "moduleResolution": "bundler",
-    "lib": ["ES2020", "DOM"],
-    "types": ["vitest/globals", "node"]
+    "module": "ESNext"
   },
   "include": ["**/*.ts"],
-  "exclude": ["node_modules", "dist"]
 }
