+++ e2e/core/vitest.config.ts
@@ -3,6 +3,9 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     globals: true,
-    environment: 'node'
+    environment: 'node',
+    typecheck: {
+      tsconfig: "./tsconfig.json"
+    }
   }
 });
