+++ packages/core/vitest.config.mjs
